Fixed bug, reason was that I missed "self" in randomforest and get_more_user_input. Also transfered Random-Forest and dim_red to the main model because its dataset-specific
